# openapi.yaml
openapi: 3.0.0
info:
  title: Person Service API
  version: 1.0.0
  description: |
    # Сервис управления данными пользователей
    
    Этот API предоставляет возможности для:
    - Управления данными пользователей (создание, обновление, удаление)
    - Управления адресами и странами
    - Работы с физическими лицами
    - Аудита изменений всех сущностей
    
    ## Основные сущности
    - Users (Пользователи)
    - Individuals (Физические лица)
    - Addresses (Адреса)
    - Countries (Страны)
    
    ## Статусы
    - ACTIVE - Активный
    - VERIFIED - Верифицированный
    - DELETED - Удаленный
    
    ## Аудит
    Все изменения сущностей отслеживаются в системе аудита.
    Доступна история изменений и просмотр конкретных ревизий.
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
paths:
...
  /api/v1/audit/users/{id}/history:
    get:
      tags:
        - audit
      summary: Получить историю изменений пользователя
      operationId: getUserHistory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешное получение истории изменений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserAuditResponse'
        '404':
          description: История изменений не найдена
  /api/v1/audit/users/{id}/revisions/{rev}:
    get:
      tags:
        - audit
      summary: Получить ревизию пользователя
      operationId: getUserRevision
      parameters:
        - $ref: '#/components/parameters/UuidParameter'
        - $ref: '#/components/parameters/RevisionParameter'
      responses:
        '200':
          description: Успешное получение ревизии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAuditResponse'
        '404':
          description: Ревизия не найдена
  /api/v1/audit/individuals/{id}/history:
    get:
      tags:
        - audit
      summary: Получить историю изменений физического лица
      operationId: getIndividualHistory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешное получение истории изменений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IndividualAuditResponse'
        '404':
          description: История изменений не найдена
  /api/v1/audit/individuals/{id}/revisions/{rev}:
    get:
      tags:
        - audit
      summary: Получить ревизию физического лица
      operationId: getIndividualRevision
      parameters:
        - $ref: '#/components/parameters/UuidParameter'
        - $ref: '#/components/parameters/RevisionParameter'
      responses:
        '200':
          description: Успешное получение ревизии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualAuditResponse'
        '404':
          description: Ревизия не найдена
  /api/v1/audit/countries/{id}:
    get:
      tags:
        - audit
      summary: Получить историю изменений страны по ID
      operationId: getCountryHistory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешное получение истории
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryAuditResponse'
        '404':
          description: Страна не найдена
  /api/v1/audit/countries/{id}/revisions/{rev}:
    get:
      tags:
        - audit
      summary: Получить ревизию страны
      operationId: getCountryRevision
      parameters:
        - $ref: '#/components/parameters/IdParameter'
        - $ref: '#/components/parameters/RevisionParameter'
      responses:
        '200':
          description: Успешное получение ревизии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryAuditResponse'
        '404':
          description: Ревизия не найдена
  /api/v1/audit/addresses/{id}/history:
    get:
      tags:
        - audit
      summary: Получить историю изменений адреса
      operationId: getAddressHistory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешное получение истории изменений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressAuditResponse'
        '404':
          description: История изменений не найдена
  /api/v1/audit/addresses/{id}/revision/{rev}:
    get:
      tags:
        - audit
      summary: Получить конкретную ревизию адреса
      operationId: getAddressRevision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rev
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешное получение ревизии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressAuditResponse'
        '404':
          description: Ревизия не найдена
components:
  parameters:
    RevisionParameter:
      name: rev
      in: path
      required: true
      schema:
        type: integer
        example: 1
    IdParameter:
      name: id
      in: path
      required: true
      schema:
        type: integer
    UuidParameter:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"
  schemas:
...
    CountryAuditResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        revisionNumber:
          type: integer
          format: int32
          description: "Номер ревизии"
        revisionType:
          type: string
          description: "Тип изменения (ADD, MOD, DEL)"
          enum: [ "ADD", "MOD", "DEL" ]
        revisionInstant:
          type: string
          format: date-time
          description: "Временная метка ревизии"
...
    AddressAuditResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        revisionNumber:
          type: integer
          format: int32
          description: "Номер ревизии"
        revisionType:
          type: string
          description: "Тип изменения (ADD, MOD, DEL)"
          enum: [ "ADD", "MOD", "DEL" ]
        revisionInstant:
          type: string
          format: date-time
          description: "Временная метка ревизии"
...
    UserAuditResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e84_uuid"
        revisionNumber:
          type: integer
          format: int32
          description: "Номер ревизии"
        revisionType:
          type: string
          description: "Тип изменения (ADD, MOD, DEL)"
          enum: [ "ADD", "MOD", "DEL" ]
        revisionInstant:
          type: string
          format: date-time
          description: "Временная метка ревизии"
...
    IndividualAuditResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        userId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        revisionNumber:
          type: integer
          format: int32
          description: "Номер ревизии"
        revisionType:
          type: string
          description: "Тип изменения (ADD, MOD, DEL)"
          enum: [ "ADD", "MOD", "DEL" ]
        revisionInstant:
          type: string
          format: date-time
          description: "Временная метка ревизии"
...